+++ src/resources/mod.rs
@@ -20,5 +20,5 @@ pub use display::{
 pub use fonts::{ExtraFonts, FontData};
 pub use lifecycle::{FadeTransition, GameState, LoadProgress, PanickedSystems, ShouldQuit};
 pub use profiling::{ProfilerData, RenderStats, SelectedEntity, SystemProfile};
-pub use render::{AmbientLight, CullConfig};
+pub use render::{AmbientLight, CullConfig, LightingConfig, DEFAULT_MAX_LIGHTS};
 pub use time::{FrameConfig, RealDt, TimeScale};